+++ src/main/kotlin/com/zepben/evolve/database/sqlite/cim/tables/iec61968/customers/TableCustomerAgreements.kt
@@ -8,8 +8,8 @@
 
 package com.zepben.evolve.database.sqlite.cim.tables.iec61968.customers
 
-import com.zepben.evolve.database.sqlite.cim.tables.Column
-import com.zepben.evolve.database.sqlite.cim.tables.Column.Nullable.NULL
+import com.zepben.evolve.database.sql.Column
+import com.zepben.evolve.database.sql.Column.Nullable.NULL
 import com.zepben.evolve.database.sqlite.cim.tables.iec61968.common.TableAgreements
 
 @Suppress("PropertyName")