+++ src/main/kotlin/com/zepben/evolve/database/sqlite/cim/tables/iec61970/base/auxiliaryequipment/TableAuxiliaryEquipment.kt
@@ -8,8 +8,8 @@
 
 package com.zepben.evolve.database.sqlite.cim.tables.iec61970.base.auxiliaryequipment
 
-import com.zepben.evolve.database.sqlite.cim.tables.Column
-import com.zepben.evolve.database.sqlite.cim.tables.Column.Nullable.NULL
+import com.zepben.evolve.database.sql.Column
+import com.zepben.evolve.database.sql.Column.Nullable.NULL
 import com.zepben.evolve.database.sqlite.cim.tables.iec61970.base.core.TableEquipment
 
 @Suppress("PropertyName")