+++ src/main/kotlin/com/zepben/evolve/database/sqlite/cim/tables/iec61970/base/core/TableBaseVoltages.kt
@@ -8,8 +8,8 @@
 
 package com.zepben.evolve.database.sqlite.cim.tables.iec61970.base.core
 
-import com.zepben.evolve.database.sqlite.cim.tables.Column
-import com.zepben.evolve.database.sqlite.cim.tables.Column.Nullable.NOT_NULL
+import com.zepben.evolve.database.sql.Column
+import com.zepben.evolve.database.sql.Column.Nullable.NOT_NULL
 
 @Suppress("PropertyName")
 class TableBaseVoltages : TableIdentifiedObjects() {