+++ src/main/kotlin/com/zepben/evolve/database/sqlite/cim/tables/iec61970/base/core/TableTerminals.kt
@@ -8,9 +8,9 @@
 
 package com.zepben.evolve.database.sqlite.cim.tables.iec61970.base.core
 
-import com.zepben.evolve.database.sqlite.cim.tables.Column
-import com.zepben.evolve.database.sqlite.cim.tables.Column.Nullable.NOT_NULL
-import com.zepben.evolve.database.sqlite.cim.tables.Column.Nullable.NULL
+import com.zepben.evolve.database.sql.Column
+import com.zepben.evolve.database.sql.Column.Nullable.NOT_NULL
+import com.zepben.evolve.database.sql.Column.Nullable.NULL
 
 @Suppress("PropertyName")
 class TableTerminals : TableAcDcTerminals() {